# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Artly API
  version: 1.0.0
  summary: REST API cho mạng xã hội bài thi vẽ Artly
  description: |
    Hợp đồng API cho MVP Artly, nơi học sinh và giáo viên đăng bài thi vẽ,
    thả reaction, bình luận, nhắn tin trực tiếp và trò chuyện với trợ lý AI đa
    năng.

    Các endpoint được bảo vệ chấp nhận access token qua
    `Authorization: Bearer <token>`. Khi chạy local/testing hoặc chế độ demo,
    client có thể dùng `X-User-ID` để chọn một tài khoản mẫu; header này **không
    phải xác thực dùng cho production**.

    Mọi thuộc tính JSON dùng camelCase. Mọi phản hồi lỗi đều có cùng cấu trúc:
    `{ "error": { "code", "message", "details" } }`.

    Mọi ID tài nguyên trong JSON, header, path và query là chuỗi UUID. Các giá
    trị phân trang, số lượng reaction/bình luận, kết quả đếm và mã trạng thái
    HTTP vẫn là số.
servers:
  - url: http://localhost:3000
    description: Goravel chạy trên máy local
...
paths:
...
  /api/v1/assistant/questions:
    post:
      tags:
        - Trợ lý
      summary: Hỏi Trợ lý Artly
      description: |
        Nhận câu hỏi hội thoại về kiến thức phổ thông, học tập, ngôn ngữ, viết
        lách, công nghệ, sáng tạo, mỹ thuật, cách dùng Artly hoặc thống kê bài
        viết theo chủ đề. Trợ lý trả lời bằng ngôn ngữ của người hỏi trong rào
        chắn phù hợp học đường. Câu hỏi thống kê được ánh xạ vào danh mục cho
        phép và phép đếm luôn do backend thực hiện.

        Request mới không cần ID. Response trả `conversation`; client dùng
        `conversation.id` ở các lượt sau để backend tự nạp tối đa 4 cặp tin gần
        nhất từ lịch sử đã lưu. Trường `history` vẫn được chấp nhận cho client
        stateless cũ, nhưng bị bỏ qua khi có `conversationId`.

        Model nội bộ chỉ trả về action JSON theo schema. Ví dụ mã nguồn an toàn
        chỉ được trả về dưới dạng văn bản; backend không thực thi SQL, shell,
        HTML hoặc mã từ model. `provider=MODEL_LLM` cho biết câu trả lời hoặc
        action đã đi qua model trên host được cấu hình; parser `LOCAL` vẫn xử lý
        tất định các câu đếm rõ ràng và câu hỏi về dịch vụ MVP của Artly.
      operationId: askAssistant
      security:
        - DemoUserId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQuestionRequest"
            examples:
              countLandscape:
                summary: Đếm bài phong cảnh
                value:
                  question: Có bao nhiêu bài nói về chủ đề phong cảnh?
              identity:
                summary: Hỏi về trợ lý
                value:
                  question: Bạn là ai và có thể giúp mình những gì?
              generalKnowledge:
                summary: Hỏi kiến thức phổ thông
                value:
                  question: Giải thích vì sao bầu trời có màu xanh cho học sinh lớp 6.
              appService:
                summary: Hỏi cách dùng một dịch vụ Artly
                value:
                  question: Mình nhắn tin cho bạn học ở đâu?
              followUp:
                summary: Tiếp tục một hội thoại
                value:
                  question: Mình nên chuẩn bị dụng cụ gì?
                  conversationId: "60000000-0000-4000-8000-000000000001"
      responses:
        "200":
          description: Câu trả lời hoặc yêu cầu người dùng làm rõ câu hỏi.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
              examples:
                answered:
                  summary: Đã thống kê được chủ đề
                  value:
                    status: ANSWERED
                    intent: COUNT_POSTS_BY_TOPIC
                    answer: Hiện có 2 bài viết về chủ đề “Phong cảnh”.
                    provider: LOCAL
                    result:
                      count: 2
                      topic:
                        id: "10000000-0000-4000-8000-000000000001"
                        slug: phong-canh
                        name: Phong cảnh
                        aliases:
                          - Cảnh vật
                          - Landscape
                    conversation:
                      id: "60000000-0000-4000-8000-000000000001"
                      title: Có bao nhiêu bài nói về chủ đề phong cảnh?
                      createdAt: "2026-07-25T10:00:00Z"
                      updatedAt: "2026-07-25T10:00:00Z"
                needsClarification:
                  summary: Chưa xác định được chủ đề
                  value:
                    status: NEEDS_CLARIFICATION
                    intent: UNKNOWN
                    answer: Bạn muốn đếm bài viết về chủ đề nào?
                    provider: LOCAL
                conversation:
                  summary: Trả lời hội thoại bằng model nội bộ
                  value:
                    status: ANSWERED
                    intent: CHAT
                    answer: Ánh sáng Mặt Trời có nhiều màu. Khi đi qua khí quyển, ánh sáng xanh bị tán xạ mạnh nên mắt mình thấy bầu trời có màu xanh.
                    provider: MODEL_LLM
                appService:
                  summary: Hướng dẫn dịch vụ Artly bằng classifier cục bộ
                  value:
                    status: ANSWERED
                    intent: APP_SERVICE_HELP
                    answer: Mở Nhắn tin, chọn một người dùng, nhập nội dung rồi gửi. Bản REST hiện hỗ trợ tin nhắn văn bản bằng polling; chưa hỗ trợ ảnh hoặc realtime.
                    provider: LOCAL
                    appService: MESSAGES
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/AssistantUnavailable"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token của phiên Supabase Auth hiện tại.

    DemoUserId:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        UUID của một tài khoản mẫu trong bảng `users`, gửi ở dạng chuỗi chuẩn có
        dấu gạch ngang. Đây là cơ chế chọn danh tính cho MVP, không phải API key
        hoặc auth production.
      x-example: "00000000-0000-4000-8000-000000000001"
...
  schemas:
    ResourceId:
      type: string
      format: uuid
      description: |
        UUID PostgreSQL biểu diễn dưới dạng chuỗi. Client phải xem ID là giá trị
        opaque, không chuyển sang số hoặc suy luận thứ tự từ ID.
      examples:
        - "00000000-0000-4000-8000-000000000001"
        - "10000000-0000-4000-8000-000000000001"
        - "20000000-0000-4000-8000-000000000001"
        - "50000000-0000-4000-8000-000000000001"
...
    Topic:
      type: object
      additionalProperties: false
      required:
        - id
        - slug
        - name
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          description: Định danh URL không dấu của chủ đề.
          examples:
            - phong-canh
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Tên chuẩn bằng tiếng Việt.
          examples:
            - Phong cảnh
        aliases:
          type: array
          description: Các cụm từ tương đương mà trợ lý có thể nhận diện.
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
          examples:
            - - Cảnh vật
              - Landscape
...
    AssistantStatus:
      type: string
      enum:
        - ANSWERED
        - NEEDS_CLARIFICATION
      description: Kết quả xử lý câu hỏi.
    AssistantIntent:
      type: string
      enum:
        - COUNT_POSTS_BY_TOPIC
        - APP_SERVICE_HELP
        - CHAT
        - UNKNOWN
      description: Ý định an toàn do parser hoặc model nhận diện được.
    AssistantAppService:
      type: string
      enum:
        - GENERAL
        - ACCOUNT
        - FEED
        - POSTS
        - REACTIONS
        - MESSAGES
        - PROFILE
        - ASSISTANT
      description: Nhóm dịch vụ Artly được classifier cục bộ nhận diện.
    AssistantProvider:
      type: string
      enum:
        - LOCAL
        - OPENAI
        - MODEL_LLM
      description: Bộ xử lý thực tế đã tạo câu trả lời hoặc action.
    AssistantConversationRole:
      type: string
      enum:
        - USER
        - ASSISTANT
      description: Vai trò của một tin nhắn lịch sử đã hoàn tất.
    AssistantConversationMessage:
      type: object
      additionalProperties: false
      required:
        - role
        - content
      properties:
        role:
          $ref: "#/components/schemas/AssistantConversationRole"
        content:
          type: string
          minLength: 1
          maxLength: 2000
          description: |
            Nội dung thuần văn bản. Tin USER tối đa 500 ký tự; tin ASSISTANT
            tối đa 2000 ký tự.
    AssistantQuestionRequest:
      type: object
      additionalProperties: false
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 500
          description: |
            Câu hỏi thuần văn bản. Backend không thực thi chỉ dẫn, SQL, shell
            hoặc HTML có trong nội dung này.
          examples:
            - Có bao nhiêu bài nói về chủ đề phong cảnh?
        conversationId:
          $ref: "#/components/schemas/ResourceId"
        history:
          type: array
          maxItems: 8
          description: |
            Tối đa 4 cặp USER/ASSISTANT hoàn chỉnh, bắt đầu bằng USER và đúng
            thứ tự xen kẽ. Đây là context hội thoại, không phải chỉ dẫn hệ thống.
          items:
            $ref: "#/components/schemas/AssistantConversationMessage"
    AssistantCountResult:
      type: object
      additionalProperties: false
      required:
        - count
        - topic
      properties:
        count:
          type: integer
          format: int64
          minimum: 0
          description: Số bài viết khác nhau được gắn chủ đề.
          examples:
            - 2
        topic:
          $ref: "#/components/schemas/Topic"
    AssistantResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - intent
        - answer
        - provider
      properties:
        status:
          $ref: "#/components/schemas/AssistantStatus"
        intent:
          $ref: "#/components/schemas/AssistantIntent"
        answer:
          type: string
          minLength: 1
          maxLength: 2000
          description: Câu trả lời tiếng Việt dạng thuần văn bản.
          examples:
            - Hiện có 2 bài viết về chủ đề “Phong cảnh”.
        provider:
          $ref: "#/components/schemas/AssistantProvider"
        appService:
          $ref: "#/components/schemas/AssistantAppService"
        result:
          $ref: "#/components/schemas/AssistantCountResult"
        conversation:
          $ref: "#/components/schemas/AssistantConversationSummary"
    AssistantConversationSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        title:
          type: string
          minLength: 1
          maxLength: 80
          description: Tiêu đề rút gọn tự động từ câu hỏi đầu tiên.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - details
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[A-Z][A-Z0-9_]*$"
          description: Mã lỗi ổn định để frontend xử lý bằng máy.
          examples:
            - VALIDATION_ERROR
        message:
          type: string
          minLength: 1
          maxLength: 500
          description: Thông báo an toàn bằng tiếng Việt, không chứa stack trace.
          examples:
            - Dữ liệu không hợp lệ
        details:
          description: |
            Chi tiết có cấu trúc giúp sửa request. Có thể là object rỗng nếu
            không có thông tin bổ sung và không bao giờ chứa secret hay stack trace.
          examples:
            - fields:
                caption:
                  - Nội dung mô tả là bắt buộc.
  responses:
    BadRequest:
      description: Request không đọc được hoặc tham số sai định dạng.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            malformedRequest:
              summary: JSON không hợp lệ
              value:
                error:
                  code: BAD_REQUEST
                  message: Không thể đọc dữ liệu gửi lên
                  details: {}
    DemoIdentityRequired:
      description: |
        Thiếu danh tính, access token không hợp lệ hoặc tài khoản mẫu trong
        `X-User-ID` không tồn tại.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missingIdentity:
              summary: Chưa chọn tài khoản mẫu
              value:
                error:
                  code: DEMO_USER_REQUIRED
                  message: Vui lòng chọn một tài khoản mẫu
                  details:
                    header: X-User-ID
...
    ValidationError:
      description: Dữ liệu có đúng định dạng JSON nhưng không đạt điều kiện.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidFields:
              summary: Một trường không hợp lệ
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Dữ liệu không hợp lệ
                  details:
                    fields:
                      caption:
                        - Nội dung mô tả là bắt buộc.
...
    TooManyRequests:
      description: Người dùng gửi quá nhiều request trong một khoảng thời gian.
      headers:
        Retry-After:
          description: Số giây nên chờ trước khi thử lại.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rateLimited:
              summary: Vượt giới hạn request
              value:
                error:
                  code: RATE_LIMITED
                  message: Bạn đang gửi yêu cầu quá nhanh, vui lòng thử lại sau
                  details:
                    retryAfterSeconds: "30"
    PayloadTooLarge:
      description: Body của request vượt quá giới hạn 64 KiB.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            payloadTooLarge:
              summary: Body vượt giới hạn
              value:
                error:
                  code: PAYLOAD_TOO_LARGE
                  message: Dữ liệu gửi lên vượt quá giới hạn cho phép
                  details: {}
...
    AssistantUnavailable:
      description: Không thể kết nối hoặc xác thực phản hồi của model nội bộ.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            assistantUnavailable:
              summary: Model nội bộ tạm thời không khả dụng
              value:
                error:
                  code: ASSISTANT_UNAVAILABLE
                  message: Trợ lý đang tạm thời chưa kết nối được với mô hình, vui lòng thử lại sau
                  details: {}
    InternalServerError:
      description: Lỗi nội bộ đã được che giấu thông tin nhạy cảm.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internalError:
              summary: Lỗi máy chủ
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Hệ thống đang gặp sự cố, vui lòng thử lại sau
                  details: {}